% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh_decomposition.R
\name{build_DFprojGram}
\alias{build_DFprojGram}
\title{Build Gram matrix for divergence-free normal projection}
\usage{
build_DFprojGram(Xe, X, Xb, Nb, kerngen)
}
\description{
Build Gram matrix for divergence-free normal projection
}
