% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_util.R
\name{calc_grad}
\alias{calc_grad}
\title{Calculate gradient of f(x,y)}
\usage{
calc_grad(xy, f, del = 1e-05)
}
\arguments{
\item{xy}{(matrix (n,2)) of points at which to calculate gradient of f}

\item{f}{(function) scalar function with two real-valued inputs}

\item{del}{(scalar) delta to use in central difference approximation of 
derivatives (see \code{central_diff()})}
}
\description{
Calculate gradient of f(x,y)
}
