% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_util.R
\name{central_diff}
\alias{central_diff}
\title{Get Central Differences}
\usage{
central_diff(f, x, del)
}
\arguments{
\item{f}{function to cacluate differences of}

\item{x}{point in domain of function at which to calculate central difference}

\item{del}{delta of central difference (f(x+0.5del) - f(x-0.5del))/2}
}
\description{
Calculate central differences (e.g. to approximate a first 
derivative of f(x))
}
