% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/vf_util.R
\name{gauss_grad}
\alias{gauss_grad}
\title{Gradient of Gaussian density}
\usage{
gauss_grad(xy, mu, sigma, del = 1e-05)
}
\arguments{
\item{mu}{(2-dim vector) bi-variate normal mean}

\item{sigma}{(matrix(2,2)) bi-variate normal covariance}

\item{del}{delta of central difference (f(x+0.5del) - f(x-0.5del))/2}
}
\description{
Gradient of Gaussian density
}
