% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_util.R
\name{get_linestrings}
\alias{get_linestrings}
\title{Get linestrings from linestring or polygon}
\usage{
get_linestrings(x)
}
\arguments{
\item{x}{linestring or polygon}
}
\description{
Get linestrings from linestring or polygon
}
