% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_util.R
\name{get_poly_normals}
\alias{get_poly_normals}
\title{Get normal vectors to a polygon}
\usage{
get_poly_normals(poly, poly_points, interior = F)
}
\arguments{
\item{poly}{polygon}

\item{poly_points}{points along polygon at which to find normal unit vectors
to the polygon boundary}

\item{interior}{(logical; default=T) should the normal vectors point to the 
interior or the exterior of the polygon}
}
\description{
Get normal vectors to a polygon
}
