\name{hhfields-package}
\alias{hhfields-package}
\alias{hhfields}
\docType{package}
\title{
\packageTitle{hhfields}
}
\description{
\packageDescription{hhfields}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{hhfields}
\packageIndices{hhfields}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{hhfields}

Maintainer: \packageMaintainer{hhfields}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
