% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hh_decomposition.R
\name{neglap_rbf2D}
\alias{neglap_rbf2D}
\title{Negative Laplacian of 2-D Radial Basis Function}
\usage{
neglap_rbf2D(X, Y, kerngen)
}
\arguments{
\item{kerngen}{kernel generator}
}
\description{
Negative Laplacian of 2-D Radial Basis Function
}
