% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_util.R
\name{rot}
\alias{rot}
\title{Rotate matrix counter clockwise through angle alpha}
\usage{
rot(angle)
}
\arguments{
\item{angle}{of rotation}
}
\value{
rotation matrix
}
\description{
Rotate matrix counter clockwise through angle alpha
}
