% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sf_util.R
\name{sample_poly_points}
\alias{sample_poly_points}
\title{Sample points on boundary of polygon}
\usage{
sample_poly_points(n, poly, remove_vert = F)
}
\arguments{
\item{n}{number of polygon points to sample}

\item{poly}{polygon list}

\item{remove_vert}{should vertices be included in sample set}
}
\description{
Sample points on boundary of polygon
}
