% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{sim_domain}
\alias{sim_domain}
\title{Simulate Domain}
\usage{
sim_domain(outer_poly, n_isq, n_bndry, n_inner)
}
\arguments{
\item{outer_poly}{(sf polygon) outer polygon.}

\item{n_isq}{number of inner squares to use}

\item{n_bndry}{Number of outer polygon boundary nodes to sample}

\item{n_inner}{Number of inner polygon boundary nodes to sample}
}
\description{
Simulate a domain over which to do a curl-free and divergence-
free dcomposition of a vector field. The domain is simulated by removing 
polygon subsets of the interior of an outer polygon. A number of interior
squares are generated and intersected with each other to define the removal 
region from the outer polygon.
}
